+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.12.0] 2025-07-14
+
+### Added
+
+- GUI_tools:
+    - New method `generate_multiselect_layout()` for the class `GUIFactory`, generates a layout for features using Listbox elements for multiple selections.
+    - `FeatureMaster` can properly handle "multi binary features"
+    - `GUIHandler` can properly handle "multi binary features"
+
 ## [3.11.0] 2025-07-14
 
 ### Added